wip: unit sharing at 'working' state, still hardcoded to share with everyone
